Move maybeEnqueueFontStyle to new helpers file actually containing helper functions
