+++ assets/src/blocks/amp-story-text/edit.js
@@ -12,29 +12,7 @@ import { RichText, InspectorControls, FontSizePicker, withFontSizes } from '@wor
 import getTagName from './get-tag-name';
 import { AMP_STORY_FONTS } from '../../constants';
 import { FontFamilyPicker } from '../../components';
-
-const maybeEnqueueFontStyle = ( slug ) => {
-	if ( ! slug || ! window.ampStoriesGoogleFonts[ slug ] ) {
-		return;
-	}
-
-	const { handle, src } = window.ampStoriesGoogleFonts[ slug ];
-
-	const element = document.getElementById( handle );
-
-	if ( element ) {
-		return;
-	}
-
-	const fontStylesheet = document.createElement( 'link' );
-	fontStylesheet.id = handle;
-	fontStylesheet.href = src;
-	fontStylesheet.rel = 'stylesheet';
-	fontStylesheet.type = 'text/css';
-	fontStylesheet.media = 'all';
-
-	document.head.appendChild( fontStylesheet );
-};
+import { maybeEnqueueFontStyle } from '../../helpers';
 
 function TextBlock( {
 	attributes,
